Deleted:AuthForm.jsx
Created:Auth module
