+++ src/components/auth/AuthForm.jsx
@@ -1,0 +1,50 @@
+import { useState } from "react";
+import { motion } from "framer-motion";
+import LoginForm from "./LoginForm";
+import SignupForm from "./SignupForm";
+import GoogleOAuth from "./GoogleOAuth";
+
+const AuthForm = () => {
+  const [isLogin, setIsLogin] = useState(true);
+
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gradient-to-r from-[#03045E] to-[#0077B6] px-4">
+      <motion.div
+        className="bg-white p-8 rounded-2xl shadow-lg w-full max-w-md"
+        initial={{ opacity: 0, y: -30 }}
+        animate={{ opacity: 1, y: 0 }}
+        transition={{ duration: 0.8 }}
+      >
+        <div className="flex justify-between mb-6">
+          <button
+            className={`text-lg font-semibold flex-1 p-2 transition-all ${
+              isLogin
+                ? "border-b-4 border-[#0077B6] text-[#0077B6]"
+                : "text-gray-500"
+            }`}
+            onClick={() => setIsLogin(true)}
+          >
+            Sign In
+          </button>
+          <button
+            className={`text-lg font-semibold flex-1 p-2 transition-all ${
+              !isLogin
+                ? "border-b-4 border-[#0077B6] text-[#0077B6]"
+                : "text-gray-500"
+            }`}
+            onClick={() => setIsLogin(false)}
+          >
+            Sign Up
+          </button>
+        </div>
+
+        {isLogin ? <LoginForm /> : <SignupForm />}
+        <div className="mt-4 flex flex-col items-center">
+          <GoogleOAuth />
+        </div>
+      </motion.div>
+    </div>
+  );
+};
+
+export default AuthForm;
